keeping common functions for SHA224 and SHA256 in same namespace
