+++ include/sha2_224.hpp
@@ -1,6 +1,5 @@
 #pragma once
-#include "utils.hpp"
-#include <CL/sycl.hpp>
+#include "sha2.hpp"
 
 namespace sha2_224 {
 
@@ -8,30 +7,10 @@ namespace sha2_224 {
 // section 1's figure 1 of Secure Hash Standard
 // http://dx.doi.org/10.6028/NIST.FIPS.180-4
 constexpr size_t IN_LEN_BITS = 448;
-constexpr size_t IN_LEN_BYTES = 56;
+constexpr size_t IN_LEN_BYTES = IN_LEN_BITS >> 3;
 
 constexpr size_t OUT_LEN_BITS = IN_LEN_BITS >> 1;
 constexpr size_t OUT_LEN_BYTES = IN_LEN_BYTES >> 1;
-
-constexpr size_t WORD_SIZE_BITS = 32;
-constexpr size_t WORD_SIZE_BYTES = 4;
-
-// SHA2-224 variant uses 64 words as constants, which are
-// specified in section 4.2.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-constexpr sycl::uint K[64] = {
-  0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1,
-  0x923f82a4, 0xab1c5ed5, 0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3,
-  0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174, 0xe49b69c1, 0xefbe4786,
-  0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
-  0x983e5152, 0xa831c66d, 0xb00327c8, 0xbf597fc7, 0xc6e00bf3, 0xd5a79147,
-  0x06ca6351, 0x14292967, 0x27b70a85, 0x2e1b2138, 0x4d2c6dfc, 0x53380d13,
-  0x650a7354, 0x766a0abb, 0x81c2c92e, 0x92722c85, 0xa2bfe8a1, 0xa81a664b,
-  0xc24b8b70, 0xc76c51a3, 0xd192e819, 0xd6990624, 0xf40e3585, 0x106aa070,
-  0x19a4c116, 0x1e376c08, 0x2748774c, 0x34b0bcb5, 0x391c0cb3, 0x4ed8aa4a,
-  0x5b9cca4f, 0x682e6ff3, 0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208,
-  0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2
-};
 
 // Initial hash values for SHA2-224, as specified in section 5.3.2 of Secure
 // Hash Standard http://dx.doi.org/10.6028/NIST.FIPS.180-4
@@ -39,58 +18,6 @@ constexpr sycl::uint IV_0[8] = {
   0xc1059ed8, 0x367cd507, 0x3070dd17, 0xf70e5939,
   0xffc00b31, 0x68581511, 0x64f98fa7, 0xbefa4fa4
 };
-
-// SHA2 function, defined in section 4.1.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-//
-// Same function is also used in SHA1
-inline sycl::uint
-ch(sycl::uint x, sycl::uint y, sycl::uint z)
-{
-  return (x & y) ^ (~x & z);
-}
-
-// SHA2 function, defined in section 4.1.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-//
-// Same function is also used in SHA1
-inline sycl::uint
-maj(sycl::uint x, sycl::uint y, sycl::uint z)
-{
-  return (x & y) ^ (x & z) ^ (y & z);
-}
-
-// SHA2 function, defined in section 4.1.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-inline sycl::uint
-Σ_0(sycl::uint x)
-{
-  return rotr(x, 2) ^ rotr(x, 13) ^ rotr(x, 22);
-}
-
-// SHA2 function, defined in section 4.1.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-inline sycl::uint
-Σ_1(sycl::uint x)
-{
-  return rotr(x, 6) ^ rotr(x, 11) ^ rotr(x, 25);
-}
-
-// SHA2 function, defined in section 4.1.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-inline sycl::uint
-σ_0(sycl::uint x)
-{
-  return rotr(x, 7) ^ rotr(x, 18) ^ (x >> 3);
-}
-
-// SHA2 function, defined in section 4.1.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-inline sycl::uint
-σ_1(sycl::uint x)
-{
-  return rotr(x, 17) ^ rotr(x, 19) ^ (x >> 10);
-}
 
 // Given 448 -bits of input ( two SHA2-224 digests concatenated ), this function
 // appends 1 -bit, required 511 -many 0 -bit and finally 64 -bits representing
@@ -171,32 +98,6 @@ pad_input_message(const sycl::uint* __restrict in,
   *(out + 31) = 0 | 0b00000001 << 8 | 0b11000000;
 }
 
-// Given 512 -bit input message block, it prepares 64 message schedules
-// for consuming input message into hash state
-//
-// See step 1 of algorithm defined in section 6.2.2 of Secure Hash Standard
-// http://dx.doi.org/10.6028/NIST.FIPS.180-4
-void
-prepare_message_schedule(const sycl::uint* __restrict in,
-                         sycl::uint* const __restrict out)
-{
-  // first 16 message schedules are same as original message words
-  // of 512 -bit message block
-#pragma unroll 16
-  for (size_t i = 0; i < 16; i++) {
-    *(out + i) = *(in + i);
-  }
-
-  // 48 iteration rounds, preparing 48 remaining message schedules
-#pragma unroll 16
-  for (size_t i = 16; i < 64; i++) {
-    const sycl::uint tmp0 = σ_1(*(out + (i - 2))) + *(out + (i - 7));
-    const sycl::uint tmp1 = σ_0(*(out + (i - 15))) + *(out + (i - 16));
-
-    *(out + i) = tmp0 + tmp1;
-  }
-}
-
 // Takes two padded, parsed input message blocks ( = 1024 -bit ) and computes
 // SHA2-224 digest ( = 224 -bit ) on input in two rounds ( because two message
 // blocks are processed sequentially )
@@ -222,7 +123,7 @@ hash(const sycl::uint* __restrict in, sycl::uint* const __restrict digest)
   for (size_t i = 0; i < 2; i++) {
     // see step 1 of algorithm defined in section  6.2.2 of Secure Hash Standard
     // http://dx.doi.org/10.6028/NIST.FIPS.180-4
-    prepare_message_schedule(in + 16 * i, msg_schld);
+    sha2::word_32::prepare_message_schedule(in + 16 * i, msg_schld);
 
     // see step 2 of algorithm defined in section  6.2.2 of Secure Hash Standard
     // http://dx.doi.org/10.6028/NIST.FIPS.180-4
@@ -241,8 +142,9 @@ hash(const sycl::uint* __restrict in, sycl::uint* const __restrict digest)
     // see step 3 of algorithm defined in section  6.2.2 of Secure Hash Standard
     // http://dx.doi.org/10.6028/NIST.FIPS.180-4
     for (size_t t = 0; t < 64; t++) {
-      sycl::uint tmp0 = h + Σ_1(e) + ch(e, f, g) + K[t] + msg_schld[t];
-      sycl::uint tmp1 = Σ_0(a) + maj(a, b, c);
+      sycl::uint tmp0 = h + sha2::word_32::Σ_1(e) + sha2::word_32::ch(e, f, g) +
+                        sha2::word_32::K[t] + msg_schld[t];
+      sycl::uint tmp1 = sha2::word_32::Σ_0(a) + sha2::word_32::maj(a, b, c);
       h = g;
       g = f;
       f = e;
